refactor sorting code to be more modular
