+++ src/common/CMakeLists.txt
@@ -3,6 +3,7 @@ add_subdirectory(enums)
 add_subdirectory(operator)
 add_subdirectory(row_operations)
 add_subdirectory(serializer)
+add_subdirectory(sort)
 add_subdirectory(types)
 add_subdirectory(value_operations)
 add_subdirectory(vector_operations)
