+++ src/common/sort/radix_sort.cpp
@@ -1,0 +1,228 @@
+#include "duckdb/common/sort/comparators.hpp"
+#include "duckdb/common/sort/sort.hpp"
+#include "duckdb/common/sort/sorted_block.hpp"
+
+namespace duckdb {
+
+//! Calls std::sort on strings that are tied by their prefix after the radix sort
+static void SortTiedBlobs(BufferManager &buffer_manager, const data_ptr_t dataptr, const idx_t &start, const idx_t &end,
+                          const idx_t &tie_col, bool *ties, const data_ptr_t blob_ptr, const SortLayout &sort_layout) {
+	const auto row_width = sort_layout.blob_layout.GetRowWidth();
+	const idx_t &col_idx = sort_layout.sorting_to_blob_col.at(tie_col);
+	// Locate the first blob row in question
+	data_ptr_t row_ptr = dataptr + start * sort_layout.entry_size;
+	data_ptr_t blob_row_ptr = blob_ptr + Load<idx_t>(row_ptr + sort_layout.comparison_size) * row_width;
+	if (!Comparators::TieIsBreakable(col_idx, blob_row_ptr, sort_layout.blob_layout)) {
+		// Quick check to see if ties can be broken
+		return;
+	}
+	// Fill pointer array for sorting
+	auto ptr_block = unique_ptr<data_ptr_t[]>(new data_ptr_t[end - start]);
+	auto entry_ptrs = (data_ptr_t *)ptr_block.get();
+	for (idx_t i = start; i < end; i++) {
+		entry_ptrs[i - start] = row_ptr;
+		row_ptr += sort_layout.entry_size;
+	}
+	// Slow pointer-based sorting
+	const int order = sort_layout.order_types[tie_col] == OrderType::DESCENDING ? -1 : 1;
+	const auto &tie_col_offset = sort_layout.blob_layout.GetOffsets()[col_idx];
+	auto logical_type = sort_layout.blob_layout.GetTypes()[col_idx];
+	std::sort(entry_ptrs, entry_ptrs + end - start,
+	          [&blob_ptr, &order, &sort_layout, &tie_col_offset, &row_width, &logical_type](const data_ptr_t l,
+	                                                                                        const data_ptr_t r) {
+		          idx_t left_idx = Load<idx_t>(l + sort_layout.comparison_size);
+		          idx_t right_idx = Load<idx_t>(r + sort_layout.comparison_size);
+		          data_ptr_t left_ptr = blob_ptr + left_idx * row_width + tie_col_offset;
+		          data_ptr_t right_ptr = blob_ptr + right_idx * row_width + tie_col_offset;
+		          return order * Comparators::CompareVal(left_ptr, right_ptr, logical_type) < 0;
+	          });
+	// Re-order
+	auto temp_block =
+	    buffer_manager.Allocate(MaxValue((end - start) * sort_layout.entry_size, (idx_t)Storage::BLOCK_SIZE));
+	data_ptr_t temp_ptr = temp_block->Ptr();
+	for (idx_t i = 0; i < end - start; i++) {
+		memcpy(temp_ptr, entry_ptrs[i], sort_layout.entry_size);
+		temp_ptr += sort_layout.entry_size;
+	}
+	memcpy(dataptr + start * sort_layout.entry_size, temp_block->Ptr(), (end - start) * sort_layout.entry_size);
+	// Determine if there are still ties (if this is not the last column)
+	if (tie_col < sort_layout.column_count - 1) {
+		data_ptr_t idx_ptr = dataptr + start * sort_layout.entry_size + sort_layout.comparison_size;
+		// Load current entry
+		data_ptr_t current_ptr = blob_ptr + Load<idx_t>(idx_ptr) * row_width + tie_col_offset;
+		for (idx_t i = 0; i < end - start - 1; i++) {
+			// Load next entry and compare
+			idx_ptr += sort_layout.entry_size;
+			data_ptr_t next_ptr = blob_ptr + Load<idx_t>(idx_ptr) * row_width + tie_col_offset;
+			ties[start + i] = Comparators::CompareVal(current_ptr, next_ptr, logical_type) == 0;
+			current_ptr = next_ptr;
+		}
+	}
+}
+
+//! Identifies sequences of rows that are tied by the prefix of a blob column, and sorts them
+static void SortTiedBlobs(BufferManager &buffer_manager, SortedBlock &sb, bool *ties, data_ptr_t dataptr,
+                          const idx_t &count, const idx_t &tie_col, const SortLayout &sort_layout) {
+	D_ASSERT(!ties[count - 1]);
+	auto &blob_block = sb.blob_sorting_data->data_blocks.back();
+	auto blob_handle = buffer_manager.Pin(blob_block.block);
+	const data_ptr_t blob_ptr = blob_handle->Ptr();
+
+	for (idx_t i = 0; i < count; i++) {
+		if (!ties[i]) {
+			continue;
+		}
+		idx_t j;
+		for (j = i; j < count; j++) {
+			if (!ties[j]) {
+				break;
+			}
+		}
+		SortTiedBlobs(buffer_manager, dataptr, i, j + 1, tie_col, ties, blob_ptr, sort_layout);
+		i = j;
+	}
+}
+
+//! Returns whether there are any 'true' values in the ties[] array
+static bool AnyTies(bool ties[], const idx_t &count) {
+	D_ASSERT(!ties[count - 1]);
+	bool any_ties = false;
+	for (idx_t i = 0; i < count - 1; i++) {
+		any_ties = any_ties || ties[i];
+	}
+	return any_ties;
+}
+
+//! Compares subsequent rows to check for ties
+static void ComputeTies(data_ptr_t dataptr, const idx_t &count, const idx_t &col_offset, const idx_t &tie_size,
+                        bool ties[], const SortLayout &sort_layout) {
+	D_ASSERT(!ties[count - 1]);
+	D_ASSERT(col_offset + tie_size <= sort_layout.comparison_size);
+	// Align dataptr
+	dataptr += col_offset;
+	for (idx_t i = 0; i < count - 1; i++) {
+		ties[i] = ties[i] && memcmp(dataptr, dataptr + sort_layout.entry_size, tie_size) == 0;
+		dataptr += sort_layout.entry_size;
+	}
+}
+
+//! Textbook LSD radix sort
+static void RadixSort(BufferManager &buffer_manager, data_ptr_t dataptr, const idx_t &count, const idx_t &col_offset,
+                      const idx_t &sorting_size, const SortLayout &sort_layout) {
+	auto temp_block = buffer_manager.Allocate(MaxValue(count * sort_layout.entry_size, (idx_t)Storage::BLOCK_SIZE));
+	data_ptr_t temp = temp_block->Ptr();
+	bool swap = false;
+
+	idx_t counts[256];
+	uint8_t byte;
+	for (idx_t offset = col_offset + sorting_size - 1; offset + 1 > col_offset; offset--) {
+		// Init counts to 0
+		memset(counts, 0, sizeof(counts));
+		// Collect counts
+		for (idx_t i = 0; i < count; i++) {
+			byte = *(dataptr + i * sort_layout.entry_size + offset);
+			counts[byte]++;
+		}
+		// Compute offsets from counts
+		for (idx_t val = 1; val < 256; val++) {
+			counts[val] = counts[val] + counts[val - 1];
+		}
+		// Re-order the data in temporary array
+		for (idx_t i = count; i > 0; i--) {
+			byte = *(dataptr + (i - 1) * sort_layout.entry_size + offset);
+			memcpy(temp + (counts[byte] - 1) * sort_layout.entry_size, dataptr + (i - 1) * sort_layout.entry_size,
+			       sort_layout.entry_size);
+			counts[byte]--;
+		}
+		std::swap(dataptr, temp);
+		swap = !swap;
+	}
+	// Move data back to original buffer (if it was swapped)
+	if (swap) {
+		memcpy(temp, dataptr, count * sort_layout.entry_size);
+	}
+}
+
+//! Identifies sequences of rows that are tied, and calls radix sort on these
+static void SubSortTiedTuples(BufferManager &buffer_manager, const data_ptr_t dataptr, const idx_t &count,
+                              const idx_t &col_offset, const idx_t &sorting_size, bool ties[],
+                              const SortLayout &sort_layout) {
+	D_ASSERT(!ties[count - 1]);
+	for (idx_t i = 0; i < count; i++) {
+		if (!ties[i]) {
+			continue;
+		}
+		idx_t j;
+		for (j = i + 1; j < count; j++) {
+			if (!ties[j]) {
+				break;
+			}
+		}
+		RadixSort(buffer_manager, dataptr + i * sort_layout.entry_size, j - i + 1, col_offset, sorting_size,
+		          sort_layout);
+		i = j;
+	}
+}
+
+void LocalSortState::SortInMemory() {
+	auto &sb = *sorted_blocks.back();
+	auto &block = sb.radix_sorting_data.back();
+	const auto &count = block.count;
+	auto handle = buffer_manager->Pin(block.block);
+	const auto dataptr = handle->Ptr();
+	// Assign an index to each row
+	data_ptr_t idx_dataptr = dataptr + sort_layout->comparison_size;
+	for (idx_t i = 0; i < count; i++) {
+		Store<idx_t>(i, idx_dataptr);
+		idx_dataptr += sort_layout->entry_size;
+	}
+	// Radix sort and break ties until no more ties, or until all columns are sorted
+	idx_t sorting_size = 0;
+	idx_t col_offset = 0;
+	unique_ptr<bool[]> ties_ptr;
+	unique_ptr<BufferHandle> ties_handle;
+	bool *ties = nullptr;
+	for (idx_t i = 0; i < sort_layout->column_count; i++) {
+		sorting_size += sort_layout->column_sizes[i];
+		if (sort_layout->constant_size[i] && i < sort_layout->column_count - 1 && sorting_size < 32) {
+			// Add columns to the sorting size until we reach a variable size column, or the last column
+			continue;
+		}
+
+		if (!ties) {
+			// This is the first sort
+			RadixSort(*buffer_manager, dataptr, count, col_offset, sorting_size, *sort_layout);
+			ties_ptr = unique_ptr<bool[]>(new bool[count]);
+			ties = ties_ptr.get();
+			std::fill_n(ties, count - 1, true);
+			ties[count - 1] = false;
+		} else {
+			// For subsequent sorts, we only have to subsort the tied tuples
+			SubSortTiedTuples(*buffer_manager, dataptr, count, col_offset, sorting_size, ties, *sort_layout);
+		}
+
+		if (sort_layout->constant_size[i] && i == sort_layout->column_count - 1) {
+			// All columns are sorted, no ties to break because last column is constant size
+			break;
+		}
+
+		ComputeTies(dataptr, count, col_offset, sorting_size, ties, *sort_layout);
+		if (!AnyTies(ties, count)) {
+			// No ties, stop sorting
+			break;
+		}
+
+		if (!sort_layout->constant_size[i]) {
+			SortTiedBlobs(*buffer_manager, sb, ties, dataptr, count, i, *sort_layout);
+			if (!AnyTies(ties, count)) {
+				// No more ties after tie-breaking, stop
+				break;
+			}
+		}
+
+		col_offset += sorting_size;
+		sorting_size = 0;
+	}
+}
+
+} // namespace duckdb
