+++ src/include/duckdb/common/types/row_data_collection.hpp
@@ -1,7 +1,7 @@
 //===----------------------------------------------------------------------===//
 //                         DuckDB
 //
-// duckdb/common/types/row_chunk.hpp
+// duckdb/common/types/row_data_collection.hpp
 //
 //
 //===----------------------------------------------------------------------===//
